Add support for decimal threshold values in alarm expressions

+++ java/monasca-common-model/src/main/java/com/hpcloud/mon/common/model/alarm/AlarmSubExpressionListener.java
@@ -143,7 +143,7 @@ class AlarmSubExpressionListener extends AlarmExpressionBaseListener {
 
   @Override
   public void exitLiteral(AlarmExpressionParser.LiteralContext ctx) {
-    threshold = Long.valueOf(ctx.getChild(0).getText());
+    threshold = Double.valueOf(ctx.getChild(0).getText());
   }
 
   @Override
